requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
ao3==0.2.0
ao3-api==2.3.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.0
certifi==2022.12.7
charset-normalizer==3.1.0
discord.py==2.2.2
frozenlist==1.3.3
idna==3.4
lxml==4.9.2
multidict==6.0.4
requests==2.28.2
soupsieve==2.4
urllib3==1.26.15
websockets==10.4
yarl==1.8.2
